過去版を削除して現行版専用に整理
