+++ README.md
@@ -2,13 +2,15 @@
 
 JRE POINT特典チケットの交換ポイントを、同じ区間・設備を現金で利用する場合の所定額と比較するWebアプリです。
 
-2026年3月14日以降の現行版と、2022年当時の普通車指定席の運賃・交換ポイントを扱う2022年版を切り替えられます。
+2026年3月14日以降の現行の運賃・料金・交換ポイントを扱います。
 
 普通車指定席では、指定席だけでなく自由席・立席の所定額とも比較できます。グリーン車とグランクラスでは、指定した設備・行程と同じ内容の所定額を比較します。
 
 通常は全乗車区間を対象に設備を選ぶだけで計算でき、必要な場合は一部区間だけ上位設備を利用する行程も指定できます。グランクラスは「飲料・軽食なしのみ」と「飲料・軽食ありを含む」を区別します。
 
-2026年版では、全線35%特別レートと、公式表の対象駅間に限った信州プレDCの交換レートも選択できます。
+全線35%特別レートと、公式表の対象駅間に限った信州プレDCの交換レートも選択できます。
+
+2022年当時の計算結果は、更新を終了した[2022年版](https://shinkansen-2022.vercel.app/)を参考資料として参照できます。現行の運賃・制度とは異なり、このリポジトリの計算ロジックには含まれません。
 
 計算対象、比較方法、区間指定の制約は[設計方針](docs/design-principles.md)にまとめています。
 
@@ -33,7 +35,7 @@ npm run build
 
 - 運賃・料金は旅客営業規則の本文と別表だけから実装します。
 - 乗換案内サイトなどが表示する金額を、計算ロジックやデータとして取り込みません。
-- `2022-03-12` と `2026-03-14` を別の年版として保持します。
+- 2026年3月14日以降の現行制度だけを実装し、過去の年版をコード内に保持しません。
 - 「はやぶさ」「こまち」の利用区間と、各座席設備の利用区間は、それぞれ連続した1区間だけを扱います。部分利用の境界駅は詳細設定で入力します。
 
 運賃・料金の一次資料は[旅客営業規則](https://www.jreast.co.jp/ryokaku/)です。交換ポイントは[えきねっとの商品ページ](https://www.eki-net.com/top/product/shinkansen/e-tokuten.html)と、そこで案内される公式表を使用します。
