+++ docs/design-principles.md
@@ -46,24 +46,21 @@
 - 秋田新幹線・山形新幹線では、グランクラスを設定できる東北新幹線内だけを入力対象とします。
 - 飛び石となる設備区間など、規則上は金額を構成できても通常のマルス計算で現実的に購入できない形は入力対象にしません。
 
-## 年版
+## 対象時点
 
-運賃・料金・交換ポイントを年版として保持し、施行日付きの規則部品を各年版から明示的に選びます。
+2026年3月14日以降の現行の運賃・料金・交換ポイントだけを計算します。過去の年版はコード内に保持せず、年度を選択する機能も設けません。
 
-- `2022-03-12`: 2022年当時のJRE POINT特典チケット（普通車指定席）の交換ポイントと所定額を扱う2022年版
-- `2026-03-14`: 2026年3月14日の運賃改定と、その時点の交換ポイントを反映した現行版
-
-古い年版は当時の比較資料として残し、現行版の規則を遡って適用しません。
+2022年版は更新終了済みの外部アーカイブとして参照できますが、現行の計算ロジック・データ・テストには含めません。
 
 ## コード構成
 
 - `App.tsx`は現在表示する画面だけを持ち、旧画面をアーカイブ目的で残しません。
-- `domain/routes.ts`は両年版で共通して使う駅・路線・営業キロと、特定都区市内等を含む区間操作を持ちます。年版固有の賃率・料金・交換ポイントは置きません。
-- `domain/seasons.ts`は両年版で共通して使うシーズン区分を持ちます。
-- `domain/basic-fares.ts`は距離帯・賃率・丸め規則から普通運賃を計算する共通処理を持ち、年版固有の金額は置きません。
-- `domain/fare-calculation.ts`は、旅客営業規則の表を検索して特急券を構成する共通処理と、施行日または確認基準日を明示した規則部品を持ちます。
-- `domain/versions/2022.ts`と`domain/versions/2026.ts`は、それぞれの年版で扱う普通運賃・シーズン規則・交換ポイントと対応機能を、同じ形の年版定義として保持します。
-- `domain/quote.ts`は年版レジストリから定義を選んで共通の見積もり処理へ渡し、年版固有の分岐やReactコンポーネント内の計算式を持ちません。
+- `domain/routes.ts`は駅・路線・営業キロと、特定都区市内等を含む区間操作を持ちます。
+- `domain/seasons.ts`は現行制度で使うシーズン区分を持ちます。
+- `domain/basic-fares.ts`は2026年3月14日施行の賃率・料金と丸め規則から普通運賃を計算します。
+- `domain/fare-calculation.ts`は、現行の旅客営業規則の表を検索して特急券を構成します。
+- `domain/current-rules.ts`は現行の交換ポイントと特別車両料金を持ちます。
+- `domain/quote.ts`は入力された行程を現行の計算処理へ渡し、Reactコンポーネント内に計算式を置きません。
 - 画面内の「区間指定」と「ランキング」はページルーティングではなくタブとして扱い、ホスティングサービス固有のrewriteを必要としない構成にします。
 
 ## 計算根拠
@@ -82,9 +79,9 @@
 
 新幹線の駅間に直接定められた特定特急料金は、旅客営業規則別表第2号の三角表を原典とします。コード上のCSVはその値を検索するための表現であり、外部の乗換案内結果から作成しません。
 
-同じ金額が複数年版に現れることだけを理由に「共通値」として扱いません。通常期の新幹線特急料金表は、2022年版アプリが当時の旅客営業規則別表第2号から転記した表を起点とし、2026年3月14日時点の同別表およびJR東日本の運送約款改正履歴と照合した規則部品として参照します。法令上の施行開始日を確認できない場合は推測せず、確認基準日と同額を確認した期限を記録します。
+通常期の新幹線特急料金表は、2026年3月14日時点の旅客営業規則別表第2号およびJR東日本の運送約款改正履歴と照合した値を使用します。法令上の施行開始日を確認できない場合は推測せず、確認基準日を記録します。
 
-シーズン加減は通常期基準額と分離します。2022年版と2026年版は、2022年4月1日施行の最繁忙期を含む4段階を使用します。根拠は旅客営業規則第57条の3、第125条およびそれらの改正履歴です。グリーン車・グランクラスについては、第125条に従いシーズン別特急料金を適用した後、530円を低減します。
+シーズン加減は通常期基準額と分離します。現行制度では、2022年4月1日施行の最繁忙期を含む4段階を使用します。根拠は旅客営業規則第57条の3、第125条およびそれらの改正履歴です。グリーン車・グランクラスについては、第125条に従いシーズン別特急料金を適用した後、530円を低減します。
 
 グランクラス(A)とグランクラス(B)を乗り継ぐ場合は、第58条第3項により全区間に対して1枚の特別車両券を発売し、第130条第2項第1号イにより全区間をグランクラス(A)として計算します。さらにグリーン車も乗り継ぐ場合は、同号ニが準用する同号ロにより、グリーン車以上の全区間に対する料金へ、グランクラス全使用区間に対するグランクラス(A)とグリーン車の差額を加えます。飲料・軽食ありの区間長だけで按分しません。
