+++ src/App.css
@@ -22,7 +22,7 @@ main { width: min(1180px, 92vw); margin: 0 auto; }
 .tabs button.active { color: #13231d; font-weight: 700; }
 .tabs button.active::after { content: ""; position: absolute; height: 4px; background: #ff5a35; left: 0; right: 0; bottom: -1px; }
 
-.control-bar { display: grid; grid-template-columns: 1.25fr 1fr 1fr; gap: 1px; background: #bcc5bd; border: 1px solid #bcc5bd; margin: 28px 0; }
+.control-bar { display: grid; grid-template-columns: 1.25fr 1fr; gap: 1px; background: #bcc5bd; border: 1px solid #bcc5bd; margin: 28px 0; }
 .control-bar label, .journey-fields label { background: #fff; padding: 13px 16px 10px; color: #65746b; font-size: 11px; font-weight: 700; }
 select { width: 100%; border: 0; border-bottom: 1px solid #94a198; background: transparent; padding: 8px 24px 8px 0; margin-top: 3px; font-size: 14px; font-weight: 700; border-radius: 0; }
 
@@ -93,6 +93,9 @@ select { width: 100%; border: 0; border-bottom: 1px solid #94a198; background: t
 
 .notice { margin: 28px 0 56px; }
 .notice p { margin: 0; color: #65746b; font-size: 12px; line-height: 1.7; }
+.site-footer { width: min(1180px, 92vw); margin: 0 auto; padding: 0 0 32px; color: #65746b; font-size: 12px; line-height: 1.7; }
+.site-footer p { margin: 0; }
+.site-footer a { color: #13231d; }
 @media (max-width: 820px) {
   .control-bar, .dashboard-grid { grid-template-columns: 1fr; }
   .journey-fields, .journey-fields.has-direction { grid-template-columns: 1fr; }
