+++ src/App.test.ts
@@ -105,11 +105,17 @@ describe("区間入力画面", () => {
     const html = renderToStaticMarkup(createElement(App));
 
     expect(html).toContain('class="journey-fields"');
-    expect(html).toContain("2022年版");
+    expect(html).not.toContain("データ年版");
+    expect(html).not.toContain("新幹線YEARスペシャル");
+    expect(html).toContain("全線35%特別レート");
     expect(html).toContain("最繁忙期");
     expect(html).toContain("乗車駅");
     expect(html).toContain("降車駅");
     expect(html).not.toContain("はやぶさ・こまち利用区間 始点");
     expect(html).not.toContain("グリーン車を利用する区間 始点");
+    expect(html).toContain(
+      'href="https://shinkansen-2022.vercel.app/">2022年版（更新終了・参考）',
+    );
+    expect(html).toContain("現行の運賃・制度とは異なります");
   });
 });
