+++ src/domain/quote.ts
@@ -1,29 +1,17 @@
-import { getBasicFareForSection, type BasicFareRules } from "./basic-fares";
 import {
-  calculateFareOptions,
-  type SeasonRules,
-  type StationExpressFareRules,
-} from "./fare-calculation";
+  getCurrentPoints,
+  getSpecialVehicleFare,
+  specialVehicleExpressReduction,
+  type Campaign,
+} from "./current-rules";
+import { calculateFareOptions } from "./fare-calculation";
 import { distanceBetween, type Line, type SortedSection } from "./routes";
 import type { Season } from "./seasons";
-import type {
-  DataVersion,
-  Facility,
-  Interval,
-  JourneySelection,
-} from "./types";
+import type { Facility, Interval, JourneySelection } from "./types";
 import { highestFacility } from "./types";
-import { version2022 } from "./versions/2022";
-import { version2026 } from "./versions/2026";
-
-export type Campaign =
-  | "regular"
-  | "shinkansenYear"
-  | "limited35Percent"
-  | "shinshuPreDc";
+export type { Campaign } from "./current-rules";
 
 export type ExclusionReason =
-  | "historicalFacility"
   | "limitedFacility"
   | "shinshuPreDc"
   | "invalidJourney";
@@ -35,44 +23,7 @@ export interface FareBreakdown {
   readonly total: number;
 }
 
-export interface FacilityDistances {
-  /** 最初にグリーン車以上へ乗ってから最後に降りるまで */
-  readonly greenKm: number;
-  /** greenKmに内包されるグランクラス区間 */
-  readonly granClassKm?: number;
-  /** A/B混在時も第130条第2項によりグランクラス全区間を(A)として計算 */
-  readonly includesGranClassA?: boolean;
-}
-
-interface PointInput {
-  readonly distanceKm: number;
-  readonly facility: Facility;
-  readonly campaign: Campaign;
-  readonly departure: string;
-  readonly arrival: string;
-}
-
-export interface VersionDefinition {
-  readonly metadata: {
-    readonly id: DataVersion;
-    readonly label: string;
-    readonly sources: Readonly<Record<string, string>>;
-    readonly note?: string;
-  };
-  readonly seasonRules: SeasonRules;
-  readonly expressFareRules: StationExpressFareRules;
-  readonly basicFareRules: BasicFareRules;
-  readonly supportsFacility: (facility: Facility) => boolean;
-  readonly getPoints: (input: PointInput) => number | undefined;
-  readonly pointExclusionReason: (campaign: Campaign) => ExclusionReason;
-  readonly specialVehicle?: {
-    readonly expressReduction: number;
-    readonly getFare: (distances: FacilityDistances) => number;
-  };
-}
-
 interface QuoteInput {
-  readonly version: DataVersion;
   readonly campaign: Campaign;
   readonly line: Line;
   readonly section: SortedSection;
@@ -97,14 +48,6 @@ interface Quote {
   readonly exclusionReason?: ExclusionReason;
 }
 
-const versions: Readonly<Record<DataVersion, VersionDefinition>> = {
-  "2022-03-12": version2022,
-  "2026-03-14": version2026,
-};
-
-export const supportedSeasonsForVersion = (version: DataVersion) =>
-  versions[version].seasonRules.supportedSeasons;
-
 const fareBreakdown = (
   basicFare: number,
   expressFare: number,
@@ -128,8 +71,7 @@ const requestedFacility = (journey: JourneySelection): Facility =>
 const ticketFare = (tickets: readonly { readonly fare: number }[]) =>
   tickets.reduce((total, ticket) => total + ticket.fare, 0);
 
-export const createQuote = ({ version: id, ...input }: QuoteInput): Quote => {
-  const version = versions[id];
+export const createQuote = (input: QuoteInput): Quote => {
   const {
     line,
     section,
@@ -166,11 +108,7 @@ export const createQuote = ({ version: id, ...input }: QuoteInput): Quote => {
       exclusionReason: "invalidJourney",
     };
   }
-  if (!version.supportsFacility(facility)) {
-    return { distanceKm, facility, exclusionReason: "historicalFacility" };
-  }
-
-  const points = version.getPoints({
+  const points = getCurrentPoints({
     distanceKm,
     facility,
     campaign: input.campaign,
@@ -181,7 +119,10 @@ export const createQuote = ({ version: id, ...input }: QuoteInput): Quote => {
     return {
       distanceKm,
       facility,
-      exclusionReason: version.pointExclusionReason(input.campaign),
+      exclusionReason:
+        input.campaign === "shinshuPreDc"
+          ? "shinshuPreDc"
+          : "limitedFacility",
     };
   }
 
@@ -192,10 +133,6 @@ export const createQuote = ({ version: id, ...input }: QuoteInput): Quote => {
       section,
       highSpeed,
       season: input.season,
-      seasonRules: version.seasonRules,
-      stationExpressFareRules: version.expressFareRules,
-      getBasicFare: (fareLine, fareSection) =>
-        getBasicFareForSection(version.basicFareRules, fareLine, fareSection),
     });
   } catch (error) {
     if (!(error instanceof RangeError)) throw error;
@@ -207,9 +144,7 @@ export const createQuote = ({ version: id, ...input }: QuoteInput): Quote => {
   }
   const expressFare =
     ticketFare(selected.expressTickets) -
-    (facility !== "ordinary"
-      ? (version.specialVehicle?.expressReduction ?? 0)
-      : 0);
+    (facility !== "ordinary" ? specialVehicleExpressReduction : 0);
 
   const intervalDistance = (interval: Interval) => {
     const start = line[interval.start];
@@ -227,14 +162,13 @@ export const createQuote = ({ version: id, ...input }: QuoteInput): Quote => {
     return { distanceKm, facility, exclusionReason: "invalidJourney" };
   }
 
-  const specialVehicleFare =
-    greenKm && version.specialVehicle
-      ? version.specialVehicle.getFare({
-          greenKm,
-          ...(granClassKm !== undefined ? { granClassKm } : {}),
-          ...(includesGranClassA ? { includesGranClassA: true } : {}),
-        })
-      : 0;
+  const specialVehicleFare = greenKm
+    ? getSpecialVehicleFare({
+        greenKm,
+        ...(granClassKm !== undefined ? { granClassKm } : {}),
+        ...(includesGranClassA ? { includesGranClassA: true } : {}),
+      })
+    : 0;
   const selectedFareBreakdown = fareBreakdown(
     selected.basicFare,
     expressFare,
